fixed not encoding image
there was an object passing problem that has been solved. The buffer object getting passed back but the buffer was getting freed before it was copied. This led to a segmentation fault that forced the program to close. The passing has been corrected, and now the image encodes and is sent back flawlessly.